Playing around actionbarsherlock

+++ mss/src/com/mss/infrastructure/data/IRepository.java
@@ -2,7 +2,7 @@ package com.mss.infrastructure.data;
 
 
 public interface IRepository<TEntity extends IEntity> {
-	TEntity getById(int id) throws Throwable;
+	TEntity getById(long id) throws Throwable;
 	Iterable<TEntity> find() throws Throwable;
 	
 	void save(TEntity entity) throws Throwable;
